// apps/api/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  PASSENGER: 'PASSENGER',
  DRIVER: 'DRIVER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const PoolStatus = {
  OPEN: 'OPEN',
  CLOSED: 'CLOSED'
} as const

export type PoolStatus = (typeof PoolStatus)[keyof typeof PoolStatus]


export const RideStatus = {
  REQUESTED: 'REQUESTED',
  MATCHED: 'MATCHED',
  ACCEPTED: 'ACCEPTED',
  DRIVER_ARRIVED: 'DRIVER_ARRIVED',
  STARTED: 'STARTED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type RideStatus = (typeof RideStatus)[keyof typeof RideStatus]